requirement.txt
amqp==5.2.0
argcomplete==3.5.1
async-timeout==4.0.3
awscli==1.35.6
bcrypt==4.2.0
billiard==4.2.0
blinker==1.8.2
boto3==1.35.13
botocore==1.35.40
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
cfn-flip==1.3.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==43.0.1
docutils==0.16
durationpy==0.9
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
hjson==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
kappa==0.6.0
kombu==5.4.1
MarkupSafe==2.1.5
placebo==0.9.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.2
redis==5.0.8
requests==2.32.3
rsa==4.7.2
s3transfer==0.10.2
six==1.16.0
SQLAlchemy==2.0.32
text-unidecode==1.3
toml==0.10.2
tqdm==4.66.5
troposphere==4.8.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uuid==1.30
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.4
zappa==0.59.0
